Support listing dual cloaks with LISTGROUPCLOAK
Dual cloaks take the form `project1/project2.foo` or
`project1/role1/project2.role2.foo` so we would like GCs of project2 to
be able to see which people have dual cloaks for their projects. We
accomplish this by finding the rightmost `/` in the cloak and then
checking for the glob `project2.*` after that point.

This also reduces the buffers needed to store these two globs, as we
know the maximum length a cloak can be, and it is a lot less than
BUFSIZE, so despite now storing two globs per cloak namespace we
overall allocate less memory than the previous implementation.