Revert "Make LISTGROUPCLOAK more efficient"
strncmp is case-sensitive whereas match is not, which makes cloaks not
exactly matching the casing in ProjectServ metadata not appear to GCs
even though those cloaks are part of their namespaces. Reverting this
commit and going back to match() is the easiest way to solve this, and
brings in IRC-specific casefolding rules in the event they are necessary
which a naive implementation would not.

This reverts commit 2fc2c8f4a19bfc72044877dde0cc0177f55d7979.