Update +S logic for single-# channels
Policy regarding single-# channels was semi-recently updated such that
informal channels are allowed there as second-class citizens to
registered project and community channels. With this change, it doesn't
make sense to override successor on single-# informal channels to
hardcode libera-placeholder-account.

Modify the logic so that we only override the successor on channels
belonging to a registered project's namespace, while leaving informal
channels to default +S behaviour.

Since this now depends on project status, the module has been moved to
the projectns/ subdirectory and given a more accurate module name for
the new functionality.