Make project successor configurable
myuser_find_ext supports ? prefixes for UID lookup and = prefixes for
currently-connected nickname lookup in addition to looking up via
account name or grouped nicks. The currently-connected nickname lookup
is a gigantic footgun in this context, however, so it is explicitly
disallowed. UIDs are the safest bet when configuring for a production
network.